requirements.txt
matplotlib==3.10.6
numpy==2.3.3
Pillow==11.3.0
torch==2.5.1+cu121
torchvision==0.20.1+cu121
